chore: reorg config code
