+++ packages/server-wallet/server-wallet.api.md
@@ -19,6 +19,7 @@ import { GetStateParams } from '@statechannels/client-api-schema';
 import { JoinChannelParams } from '@statechannels/client-api-schema';
 import { JSONSchema } from 'objection';
 import Knex from 'knex';
+import { Level } from 'pino';
 import { Logger } from 'pino';
 import { Payload as Message } from '@statechannels/wallet-core';
 import { MessageQueuedNotification } from '@statechannels/client-api-schema';
@@ -30,7 +31,6 @@ import { Outcome } from '@statechannels/wallet-core';
 import { Participant } from '@statechannels/wallet-core';
 import { Participant as Participant_2 } from '@statechannels/client-api-schema';
 import { Payload } from '@statechannels/wire-format';
-import * as pino from 'pino';
 import { Pojo } from 'objection';
 import { PrivateKey } from '@statechannels/wallet-core';
 import { providers } from 'ethers';
