+++ packages/server-wallet/src/config/index.ts
@@ -1,0 +1,3 @@
+export * from './types';
+export * from './utils';
+export * from './defaults';
