+++ packages/server-wallet/src/wallet/__test__/integration/sync-channel.test.ts
@@ -6,10 +6,10 @@ import {seedAlicesSigningWallet} from '../../../db/seeds/1_signing_wallet_seeds'
 import {stateWithHashSignedBy} from '../fixtures/states';
 import {alice, bob, charlie} from '../fixtures/signing-wallets';
 import * as participantFixtures from '../fixtures/participants';
-import {channel} from '../../../models/__test__/fixtures/channel';
 import {testKnex as knex} from '../../../../jest/knex-setup-teardown';
 import {defaultTestConfig} from '../../../config';
 import {DBAdmin} from '../../../db-admin/db-admin';
+import {channel} from '../../../models/__test__/fixtures/channel';
 
 let w: Wallet;
 beforeEach(async () => {
